+++ game_patch/rf/network.h
@@ -7,6 +7,7 @@ namespace rf
 {
     // Forward declarations
     class Player;
+    class EntityObj;
 
     struct NwAddr
     {
@@ -106,7 +107,8 @@ namespace rf
     static auto& BanIp = AddrAsRef<void(const NwAddr& addr)>(0x0046D0F0);
     static auto& MultiSetRequestedWeapon = AddrAsRef<void(int weapon_cls_id)>(0x0047FCA0);
     static auto& MultiChangeLevel = AddrAsRef<void(const char* filename)>(0x0047BF50);
-    static auto& PingPlayer = AddrAsRef<void(rf::Player*)>(0x00484D00);
+    static auto& PingPlayer = AddrAsRef<void(Player*)>(0x00484D00);
+    static auto& SendEntityCreatePacket = AddrAsRef<void(EntityObj *entity, Player* player)>(0x00475160);
 
     static auto& serv_addr = AddrAsRef<NwAddr>(0x0064EC5C);
     static auto& serv_name = AddrAsRef<String>(0x0064EC28);