+++ game_patch/server/server.h
@@ -11,3 +11,4 @@ void ServerCleanup();
 void ServerDoFrame();
 bool CheckServerChatCommand(const char* msg, rf::Player* sender);
 void ServerOnLimboStateEnter();
+bool ServerIsSavingEnabled();